Adding WeaponComponent.h and .cpp Updating FPDevCharacter.h and .cpp
Created a new USkeletalMeshComponent subclass WeaponComponent.
WeaponComponent now holds FiringSound, FiringAnimation, and Projectile
Class members that were part of the First Person Shooter Template.

FPDevChatacter now has a method for attaching weapons
AttachWeapon(UClass* CompClass).  This method registers a new componet
in the weapons slot FP_GUN.
TODO:
This needs to be updated to clear out and unregister the component in
FP_Gun and create a new component, or manage a list of weapon component.