SpreadPattern added, MortalInterface and subclasses updated.
Minor adjustmens to the MortalInterface and classes that use it.

FPDevCharacter class updates to used SpreadPattern functionality of
WeaponMechanic class.  This is still a work is progress.

Basic WeaponSpread for a single shot shot must use a width of 3 with
second index true.  This is to compensate for some math problems.