BugFix and Improvement for Weapon FireDelay
Cleaned up the code for weapon fireing in FPDevCharacter.  In the
process fixed a bug where TimeSinceFire was not getting reset by OnFire
when a "bullet" was added to FireQue.  This mean FireDelay was not
working properly but is now.

As an Improvement to WeaponMechanics a new boolean has been added called
FullAutomatic, when FullAutomic is ture the mouse can be held to repeat
firing at FireDelay rate.   FullAutomaic can be modified with the
helpfer function ToggleFullAutomatic()

I really need to stop messing with Weapons at this point and move on to
other things... after updating the weapons documentation... lol