WeaponMechanic Improvments
WeaponMechanic impromevments included prevent weapons from "queuing"
rounds when the user clicks more often than the fire multiplier.

In other words the weapon stops firing when the player stops shooting,
and does not fire more often than once per FireDelay, no matter how fast
the player clicks.

Minimum MultiplierDelay is now 0.001.  ModiflyMultiplierDelay checks and
sets to this minimum if Delay give is invalid.

ModifyFireDelay(Delay) now checks that FireDelay is no less than
MultiplierDelay * ShotMultiplier, and uses this value as a minimum
otherwise.

Changing the ShotMultiplierDelay now resets FireDelay with a
ModifyFireDelay call checking and reseting FireDelay to the minimum if
it would be invalid.

Changing the ShotMultiplier now also resets FireDelay with a
ModifyFireDelay call checking and reseting FireDelay to the minimum if
it would be invalid.