In Progress refactor.
This is an overhual/refactor of the project.

Changes in this refactor are focused on the heirarchy of gameplay
classes which previously inherited the IMortalInterface health system.
Health mechanics are now components which the primary gameplay framework
classes MortalActor, MortalPawn, and MortalCharacter (subclasses of
Unreal's Actor, Pawn, and Character)  are now composed of.

Also included  in this refactor are changes to existing compoents like
WeaponMechanic, WeaponComponent, and ImpulseEngine.  More information
about changes will be documented in the wiki.

Redundant classes have been removed like IWeaponInterface, and
ProjectileInterface.  The baseclasses for a heirarchy will be enough to
define the interface for these types of objects.  Other classes that are
currently removed AI and AI Mortal base classes, and HUD UI classes will
be added as they are redeveloped.

The major reason for branching this refactor from the master is to allow
time for me to update  all the documentation, prior to releaseing the
refactored framework.  The Project name will eventually change to
reflect the name used in the refactored classes.